fix: added common eslint configuration
