+++ package.json
@@ -8,10 +8,8 @@
   "devDependencies": {
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-conventional": "^11.0.0",
-    "@typescript-eslint/eslint-plugin": "^4.23.0",
-    "@typescript-eslint/parser": "^4.23.0",
     "@voiceflow/commitlint-config": "^1.0.1",
-    "@voiceflow/eslint-config": "^2.0.5",
+    "@voiceflow/eslint-config": "^3.1.0",
     "@voiceflow/git-branch-check": "^1.1.3",
     "@voiceflow/prettier-config": "^1.0.5",
     "commitizen": "^4.2.3",
@@ -20,7 +18,7 @@
     "fixpack": "^4.0.0",
     "husky": "^4.3.8",
     "lerna": "^4.0.0",
-    "lint-staged": "^10.5.3",
+    "lint-staged": "^11.0.0",
     "prettier": "^2.3.0",
     "prettier-eslint-cli": "^5.0.0",
     "rimraf": "^3.0.0"
