+++ packages/alexa-types/lib/constants/slots.ts
@@ -1,3 +1,4 @@
+/* eslint no-secrets/no-secrets: 0 */
 import { BuiltinSlot } from '@voiceflow/common';
 
 import { Locale as L } from '@/types';
