+++ packages/alexa-types/lib/index.ts
@@ -1,8 +1,8 @@
-export * from './types';
-export * from './version';
 export * from './constants';
 export * from './diagram';
-export * from './program';
 export * from './nodes';
-export * from './utils';
+export * from './program';
 export * from './project';
+export * from './types';
+export * from './utils';
+export * from './version';
