+++ packages/alexa-types/lib/nodes/accountLinking.ts
@@ -1,3 +1,5 @@
+/* eslint-disable camelcase */
+
 import { NodeID } from '@voiceflow/general-types';
 
 import { DefaultNode, DefaultStep, NodeType } from './types';
