+++ packages/alexa-types/package.json
@@ -15,6 +15,7 @@
   "devDependencies": {
     "@types/compression": "^1.0.1",
     "@types/lodash": "^4.14.149",
+    "@voiceflow/tsconfig": "^1.0.0",
     "@zerollup/ts-transform-paths": "^1.7.9",
     "depcheck": "^1.3.1",
     "rimraf": "^3.0.0",
