+++ packages/api-sdk/package.json
@@ -19,6 +19,7 @@
     "@types/jsonwebtoken": "^8.5.0",
     "@types/mocha": "^8.0.0",
     "@types/sinon": "^9.0.4",
+    "@voiceflow/tsconfig": "^1.0.0",
     "@zerollup/ts-transform-paths": "^1.7.18",
     "chai": "^4.2.0",
     "depcheck": "^1.3.1",
@@ -47,7 +48,7 @@
   "scripts": {
     "build": "npm run clean && ttsc --project ./tsconfig.build.json",
     "clean": "rimraf build",
-    "lint": "eslint \"src/**/*.{ts,js}\"",
+    "lint": "eslint \"src/**/*.{ts,js}\" \"tests/**/*.{ts,js}\"",
     "lint:fix": "yarn lint --fix",
     "lint:quiet": "yarn lint --quiet",
     "lint:report": "yarn lint --format json -o sonar/report.json",
