+++ packages/api-sdk/src/models/index.ts
@@ -1,7 +1,7 @@
-export * from './shared';
-export * from './member';
 export * from './apiKey';
-export * from './version';
-export * from './project';
-export * from './program';
 export * from './diagram';
+export * from './member';
+export * from './program';
+export * from './project';
+export * from './shared';
+export * from './version';
