+++ packages/api-sdk/src/resources/index.ts
@@ -1,7 +1,7 @@
-export { default as User } from './user';
 export { default as APIKey } from './apiKey';
-export { default as Project } from './project';
-export { default as Version } from './version';
-export { default as Program } from './program';
 export { default as Diagram } from './diagram';
+export { default as Program } from './program';
+export { default as Project } from './project';
 export { default as PrototypeProgram } from './prototypeProgram';
+export { default as User } from './user';
+export { default as Version } from './version';
