+++ packages/api-sdk/src/resources/project/constants.ts
@@ -1,2 +1,1 @@
-// eslint-disable-next-line import/prefer-default-export
 export const ENDPOINT = 'projects';
