+++ packages/general-types/lib/config/index.ts
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/prefer-default-export
 export class Config {
   tts?: boolean;
