+++ packages/general-types/lib/constants/index.ts
@@ -1,4 +1,4 @@
-export * from './slot';
-export * from './intent';
-export * from './device';
 export * from './azureVoices';
+export * from './device';
+export * from './intent';
+export * from './slot';
