+++ packages/general-types/lib/index.ts
@@ -1,9 +1,9 @@
-export * from './types';
-export * from './nodes';
-export * from './project';
-export * from './version';
-export * from './request';
 export * from './config';
 export * from './constants';
+export * from './nodes';
+export * from './project';
+export * from './request';
 export * from './trace';
+export * from './types';
+export * from './version';
 export * from './vf';
