+++ packages/general-types/package.json
@@ -10,6 +10,7 @@
     "@voiceflow/api-sdk": "^1.31.6"
   },
   "devDependencies": {
+    "@voiceflow/tsconfig": "^1.0.0",
     "@zerollup/ts-transform-paths": "^1.7.9",
     "depcheck": "^1.3.1",
     "rimraf": "^3.0.0",
