+++ packages/google-types/lib/constants/index.ts
@@ -1,5 +1,5 @@
-export * from './intents';
 export * from './configFile';
+export * from './intents';
 export * from './locales';
 export * from './slots';
 export * from './voices';
